updated models, added additional checks in dataHelper.loadData method

+++ src/main/java/com/mobiquityinc/model/Thing.java
@@ -3,13 +3,13 @@ package com.mobiquityinc.model;
 public class Thing {
     private int index;
     private double weight;
-    private double cost;
+    private int cost;
 
     public Thing() {
 
     }
 
-    public Thing(int index, double weight, double cost) {
+    public Thing(int index, double weight, int cost) {
         this.index = index;
         this.weight = weight;
         this.cost = cost;
@@ -31,11 +31,11 @@ public class Thing {
         this.weight = weight;
     }
 
-    public double getCost() {
+    public int getCost() {
         return cost;
     }
 
-    public void setCost(double cost) {
+    public void setCost(int cost) {
         this.cost = cost;
     }
 }
